Images purge, Pseudo Cron, Detail route changed

+++ app/models/ImageModel.php
@@ -6,11 +6,12 @@ namespace App\Model;
 use App\Model\Entities\Image;
 use App\Model\Entities\User;
 use Nette\Http\FileUpload;
+use Tracy\Debugger;
 
 class ImageModel extends BaseModel
 {
 
-    public function createImage(FileUpload $imageData, $userId)
+    public function createImage(FileUpload $imageData)
     {
 
         $name = uniqid("img_");
@@ -26,13 +27,11 @@ class ImageModel extends BaseModel
                 $extension = "gif";
                 break;
         }
-
-        $user = $this->getOne(User::class, array("id" => $userId));
+        
 
         $target = "assets/images/uploads/" . $name . "." . $extension;
 
         $image = new Image();
-        $image->setUser($user);
         $image->setName($name);
         $image->setExtension($extension);
 
@@ -48,19 +47,61 @@ class ImageModel extends BaseModel
 
     public function deleteImage($id)
     {
-        $image = $this->getOne(Image::class, array("id" => $id));
 
-        if ($image) {
-            $target = "assets/images/uploads/" . $image->getName() . "." . $image->getExtension();
+        if (is_array($id)) {
+            foreach ($id as $imageId) {
+                $image = $this->getOne(Image::class, array("id" => $imageId));
 
-            unlink($target);
+                if ($image) {
+                    $target = "assets/images/uploads/" . $image->getName() . "." . $image->getExtension();
 
-            if (!file_exists($target)) {
-                $this->remove($image);
-                $this->flush();
+                    if(file_exists($target)){
+                        unlink($target);
+                    }
+
+                    if (!file_exists($target)) {
+                        $this->remove($image);
+                    }
+                }
+            }
+        } else {
+            $image = $this->getOne(Image::class, array("id" => $id));
+
+            if ($image) {
+                $target = "assets/images/uploads/" . $image->getName() . "." . $image->getExtension();
+
+                if(file_exists($target)){
+                    unlink($target);
+                }
+
+                if (!file_exists($target)) {
+                    $this->remove($image);
+                }
             }
         }
 
+        $this->flush();
+    }
+
+    public function purgeImages()
+    {
+        $imageIds = array();
+        $time = new \DateTime();
+        $time->modify("-1 day");
+
+        $query = $this->getEm()->createQuery('
+            SELECT i
+            FROM App\Model\Entities\Image i
+            WHERE i.tutorial is NULL
+            AND i.upDate < :time'
+        )->setParameter('time', $time);
+        $images = $query->getResult();
+
+        foreach ($images as $image) {
+            $imageIds[] = $image->getId();
+        }
+
+        $this->deleteImage($imageIds);
     }
     
 }
